shared/bus-util: move bus_message_read_id128() to bus-message-util
(cherry picked from commit ab33edb05b7d4c90fb80f46aa6b951c505048798)

Resolves: RHEL-73780
