+++ src/shared/bus-message-util.h
@@ -6,6 +6,8 @@
 #include "in-addr-util.h"
 #include "socket-netlink.h"
 
+int bus_message_read_id128(sd_bus_message *m, sd_id128_t *ret);
+
 int bus_message_read_ifindex(sd_bus_message *message, sd_bus_error *error, int *ret);
 int bus_message_read_family(sd_bus_message *message, sd_bus_error *error, int *ret);
 int bus_message_read_in_addr_auto(sd_bus_message *message, sd_bus_error *error, int *ret_family, union in_addr_union *ret_addr);
