+++ src/shared/bus-util.c
@@ -953,34 +953,6 @@ int bus_query_sender_pidref(
         return bus_creds_get_pidref(creds, ret);
 }
 
-int bus_message_read_id128(sd_bus_message *m, sd_id128_t *ret) {
-        const void *a;
-        size_t sz;
-        int r;
-
-        assert(m);
-
-        r = sd_bus_message_read_array(m, 'y', &a, &sz);
-        if (r < 0)
-                return r;
-
-        switch (sz) {
-        case 0:
-                if (ret)
-                        *ret = SD_ID128_NULL;
-                return 0;
-
-        case sizeof(sd_id128_t):
-                if (ret)
-                        memcpy(ret, a, sz);
-                return !memeqzero(a, sz); /* This mimics sd_id128_is_null(), but ret may be NULL,
-                                           * and a may be misaligned, so use memeqzero() here. */
-
-        default:
-                return -EINVAL;
-        }
-}
-
 static const char* const bus_transport_table[] = {
         [BUS_TRANSPORT_LOCAL]   = "local",
         [BUS_TRANSPORT_REMOTE]  = "remote",
