+++ src/shared/bus-util.h
@@ -84,6 +84,4 @@ int bus_property_get_string_set(sd_bus *bus, const char *path, const char *inter
 int bus_creds_get_pidref(sd_bus_creds *c, PidRef *ret);
 int bus_query_sender_pidref(sd_bus_message *m, PidRef *ret);
 
-int bus_message_read_id128(sd_bus_message *m, sd_id128_t *ret);
-
 const char* bus_transport_to_string(BusTransport transport) _const_;
